% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{cleandata}
\alias{cleandata}
\title{Clean library circulation data.}
\usage{
cleandata(librarydata)
}
\arguments{
\item{librarydata}{Data frame containing at least three columns labeled
CALLNUM, PAGES, and PRICE.}
}
\value{
Returns a data frame with numeric values for CALLNUM, PAGES, and
  PRICE and items with invalid entries removed.
}
\description{
\code{cleandata} converts and cleans raw data obtained from library
management systems (i.e SIRSI), in preparation for use in regression
modeling.
}
\details{
This function accepts a data frame containing bibliographic data for
non-fiction monographs using Dewey Decimal Classification and must contain at
least the following named columns: CALLNUM - a string consisting of
alpha-numeric call numbers; PAGES - string containing the number of pages in
the monograph (e.g. MARC tag for physical description);PRICE - a string
containing the price of the item.The function extracts numeric values for
dewey classification, number of pages, and price for each item. It then
removes all items with CALLNUM not in the range of 0 to 1000, all items with
values for PAGES less than 5 or NA, and all items with PRICE equal to zero.
}
\examples{
cleandata(libdata)
}
