% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deweyhistograms.R
\name{deweygraph}
\alias{deweygraph}
\title{Graph Circulation By Dewey Classification}
\usage{
deweygraph(librarydata)
}
\arguments{
\item{librarydata}{Data frame which must contain at least two columns, one labeled
CHARGES containing integer data, another labeled CALLNUM containing numeric data.}
}
\value{
Returns a bar graph.
}
\description{
\code{deweygraph} produces a bar graph showing the total aggregate
circulation for 10 broad dewey classifications.
}
\details{
This function accepts a data frame containing bibliographic data for
non-fiction monographs using Dewey Decimal Classification and must contain at
least the following named columns: CHARGES - an integer for the total number
of charges for each item; CALLNUM - a numeric value describing the item's
Dewey Decimal Classification. A bar graph is produced using ggplot2,
subdividing data according to 10 broad Dewey Decimal Classes (100s, 200s,
etc.) and showing total aggregate circulation for all items in each class.
}
\examples{
deweygraph(cleanlib)

}
